(svn r3567) -Fix: for once and for all, use 'unsigned long' if you want to cast something from int-like to void *... we fixed that not too long ago, and the code below this changed line clearly tells you do to that :p

+++ settings.c
@@ -550,7 +550,7 @@ static const void *string_to_val(const SettingDesc *desc, const char *str)
 	switch (GetSettingGenericType(desc)) {
 	case SDT_NUMX: {
 		char *end;
-		uint32 val = strtoul(str, &end, 0);
+		unsigned long val = strtoul(str, &end, 0);
 		if (*end != '\0') ShowInfoF("ini: trailing characters at end of setting '%s'", desc->name);
 		return (void*)val;
 	}
